Refactor: move diagrams to separate module
