+++ lib/jekyll-diagrams.rb
@@ -5,56 +5,6 @@ require 'jekyll'
 module Jekyll
   module Diagrams
     class << self
-      # Return configuration of Jekyll Diagrams
-      #
-      # @param context [Liquid::Context, :registers] Parsed context
-      # @return Configuration
-      def configuration(context)
-        site_config = context.registers[:site].config
-        page_config = context.registers[:page]
-
-        site_config.merge(page_config)
-      end
-
-      def config_for(context, name)
-        configuration(context).dig(config_name, name) || {}
-      end
-
-      def error_mode(context)
-        liquid_mode = configuration(context).dig('liquid', 'error_mode')
-        custom_mode = configuration(context).dig(config_name, 'error_mode')
-
-        (custom_mode || liquid_mode || :warn).to_sym
-      end
-
-      # Return file path under vendor path
-      #
-      # @param file [String] If not given, return directory path
-      def vendor_path(file = '')
-        File.join(File.expand_path('../vendor', __dir__), file)
-      end
-
-      # @param jar [String] Jar path to run
-      def run_jar(jar)
-        "java -Djava.awt.headless=true -jar #{jar} "
-      end
-
-      def normalized_attrs(attrs, prefix:, sep: '=')
-        attrs =
-          case attrs
-          when String
-            attrs
-          when Array
-            attrs.join(prefix)
-          when Hash
-            attrs.map { |k, v| "#{k}#{sep}#{v}" }.join(prefix)
-          end
-
-        "#{prefix}#{attrs}"
-      end
-
-      private
-
       def config_name
         'jekyll-diagrams'
       end
@@ -63,8 +13,9 @@ module Jekyll
 end
 
 require_relative 'jekyll-diagrams/errors'
+require_relative 'jekyll-diagrams/utils'
 require_relative 'jekyll-diagrams/rendering'
-require_relative 'jekyll-diagrams/block'
+require_relative 'jekyll-diagrams/renderer'
 
 require_relative 'jekyll-diagrams/blockdiag'
 require_relative 'jekyll-diagrams/erd'
