+++ lib/jekyll-diagrams/smcat/block.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Jekyll
+  module Diagrams
+    module SMCat
+      class Block < Liquid::Block
+        def render(context)
+          Renderer.render(context, super.to_s, block_name)
+        end
+      end
+    end
+  end
+end
