+++ lib/jekyll-diagrams/svgbob.rb
@@ -1,27 +1,12 @@
 # frozen_string_literal: true
 
+require_relative 'svgbob/renderer'
+require_relative 'svgbob/block'
+
 module Jekyll
   module Diagrams
-    class SvgbobBlock < Block
-      CONFIGURATIONS = %w[font-family font-size scale stroke-width].freeze
-
-      def render_svg(code, config)
-        command = build_command(config)
-
-        render_with_stdin_stdout(command, code)
-      end
-
-      def build_command(config)
-        command = +'svgbob'
-
-        CONFIGURATIONS.each do |conf|
-          command << " --#{conf} #{config[conf]}" if config.key?(conf)
-        end
-
-        command
-      end
+    module Svgbob
+      Liquid::Template.register_tag(:svgbob, Block)
     end
   end
 end
-
-Liquid::Template.register_tag(:svgbob, Jekyll::Diagrams::SvgbobBlock)
