+++ lib/jekyll-diagrams/svgbob/renderer.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+module Jekyll
+  module Diagrams
+    module Svgbob
+      class Renderer < BaseRenderer
+        CONFIGURATIONS = %w[font-family font-size scale stroke-width].freeze
+
+        def render_svg(code, config)
+          command = build_command(config)
+
+          render_with_stdin_stdout(command, code)
+        end
+
+        def build_command(config)
+          command = +'svgbob'
+
+          CONFIGURATIONS.each do |conf|
+            command << " --#{conf} #{config[conf]}" if config.key?(conf)
+          end
+
+          command
+        end
+      end
+    end
+  end
+end
