+++ lib/jekyll-diagrams/syntrax.rb
@@ -1,29 +1,12 @@
 # frozen_string_literal: true
 
+require_relative 'syntrax/renderer'
+require_relative 'syntrax/block'
+
 module Jekyll
   module Diagrams
-    class SyntraxBlock < Block
-      CONFIGURATIONS = %w[scale style].freeze
-
-      def render_svg(code, config)
-        command = build_command(config)
-
-        render_with_tempfile(command, code) do |input, output|
-          "--input #{input} --output #{output}"
-        end
-      end
-
-      def build_command(config)
-        command = +'syntrax'
-
-        CONFIGURATIONS.each do |conf|
-          command << " --#{conf} #{config[conf]}" if config.key?(conf)
-        end
-
-        command
-      end
+    module Syntrax
+      Liquid::Template.register_tag(:syntrax, Block)
     end
   end
 end
-
-Liquid::Template.register_tag(:syntrax, Jekyll::Diagrams::SyntraxBlock)
