+++ lib/jekyll-diagrams/wavedrom.rb
@@ -1,21 +1,12 @@
 # frozen_string_literal: true
 
+require_relative 'wavedrom/renderer'
+require_relative 'wavedrom/block'
+
 module Jekyll
   module Diagrams
-    class WavedromBlock < Block
-      def render_svg(code, config)
-        command = build_command(config)
-
-        render_with_tempfile(command, code) do |input, output|
-          "--input #{input} --svg #{output}"
-        end
-      end
-
-      def build_command(_config)
-        'wavedrom-cli'
-      end
+    module Wavedrom
+      Liquid::Template.register_tag(:wavedrom, Block)
     end
   end
 end
-
-Liquid::Template.register_tag(:wavedrom, Jekyll::Diagrams::WavedromBlock)
