+++ test/svgbob_test.rb
@@ -4,19 +4,18 @@ require 'test_helper'
 
 class SvgbobTest < Minitest::Test
   def setup
-    @block = Jekyll::Diagrams::SvgbobBlock.parse(
-      'svgbob',
+    @renderer = Jekyll::Diagrams::Svgbob::Renderer.new(
+      Liquid::ParseContext.new,
       '',
-      Liquid::Tokenizer.new('test {% endsvgbob %}'),
-      Liquid::ParseContext.new
+      'svgbob'
     )
   end
 
   def test_build_command_with_default_config
     config = {}
-    assert_equal 'svgbob', @block.build_command(config)
+    assert_equal 'svgbob', @renderer.build_command(config)
 
     config = { 'scale' => '2' }
-    assert_match '--scale 2', @block.build_command(config)
+    assert_match '--scale 2', @renderer.build_command(config)
   end
 end
